Move joystick polling into its own thread
Use GThread and GMutex to start and stop polling in a separate thread.
Currently the only UI interaction is starting polling by selecting a
device and stopping polling via a button, and putting message on the
statusbar about starting/stopping, actual reporting of events in the UI
needs to be implemented.